Refactor neighbors API and introduce the meta neighbors estimators
