+++ sktree/__init__.py
@@ -36,7 +36,7 @@ if __sktree_SETUP__:
     # process, as it may not be compiled yet
 else:
     try:
-        from . import _lib, tree, ensemble, experimental
+        from . import _lib, tree, ensemble, experimental, neighbors
         from .ensemble._unsupervised_forest import (
             UnsupervisedRandomForest,
             UnsupervisedObliqueRandomForest,
@@ -59,6 +59,7 @@ else:
         "tree",
         "experimental",
         "ensemble",
+        "neighbors",
         "ObliqueRandomForestClassifier",
         "ObliqueRandomForestRegressor",
         "PatchObliqueRandomForestClassifier",
