+++ sktree/tree/meson.build
@@ -16,7 +16,6 @@ foreach ext: extensions
   )
 endforeach
 
-# TODO: comment in _classes.py when we have a working Cython unsupervised tree with a Python API
 python_sources = [
   '__init__.py',
   '_classes.py',
